pyproject.toml
# Copyright 2022 simple-syslog authors
# All rights reserved.
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#       http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
[tool.poetry]
name = "simple-syslog"
version = "0.1.0"
license = "Apache-2.0"
description = "A python 3 library for parsing valid RFC 5324 and RFC 3164 syslog"
authors = ["Otto Fowler <ottobackwards@gmail.com>"]
maintainers = ["palindromicity <github.palindromicity@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/palindromicity/simple-syslog-python"
repository = "https://github.com/palindromicity/simple-syslog-python"
keywords = ["syslog", "RFC 5424", "RFC 3164", "simple-syslog"]

[tool.poetry.urls]
issues = "https://github.com/palindromicity/simple-syslog-python/issues"

[tool.coverage.paths]
source = ["simple_syslog", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["simple_syslog"]

[tool.coverage.report]
show_missing = true

#[tool.coverage.report]
#fail_under = 100
[tool.isort]
profile = "black"

[tool.mypy]
warn_return_any = true
exclude = "simple_syslog/generated"
[[tool.mypy.overrides]]
module = "nox.*"
ignore_missing_imports = true
ignore_errors = true
[[tool.mypy.overrides]]
module = "noxfile"
ignore_missing_imports = true
ignore_errors = true
[[tool.mypy.overrides]]
module = "simple_syslog.generated.grammars.*"
ignore_missing_imports = true
ignore_errors = true
[[tool.mypy.overrides]]
module = "antlr4.*"
ignore_missing_imports = true
ignore_errors = true

[tool.poetry.dependencies]
python = "^3.8"
antlr4-python3-runtime = "^4.10"
black = "^22.6.0"
flake8 = "^5.0.4"
flake8-black = "^0.3.3"
types-flake8-bugbear = "^22.7.1"
isort = "^5.10.1"
flake8-isort = "^4.2.0"
flake8-bugbear = "^22.7.1"
autoflake = "^1.4"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
coverage = {extras = ["toml"], version = "^6.4.4"}
pytest-cov = "^3.0.0"
pytest-mock = "^3.8.2"
mypy = "^0.971"
flake8-docstrings = "^1.6.0"
black = "^22.6.0"
darglint = "^1.8.1"
xdoctest = "^1.0.2"
nox = "^2022.8.7"
nox-poetry = "^1.0.1"
Sphinx = "^5.1.1"
sphinx-autodoc-typehints = "^1.19.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
